Cleanup and Case sensitive sorting
Cleaned up the listener class and driver class while also modifying the datamanager class to have a fileName variable so that we can specifiy the fileName uniquely. This will allow multiple files to be utilized should we desire and we could even implement a gui button that lets you specify a file to make 'active' and then pull from so that the list uses that file and when things are added they add to said file. Also mondified the multlist name sort slightly to accompadate for case sensitive names. Previously all lower case would be at the end of the list, now lowercase come after uppercase but still in alphabetical order. Required a bit of tedious code to make work as the process of ensuring they always follow that pattern is a tad more complicated than it sounds. Added more comments and set up the driver so that all classes are initialized and then linked and the file is read. Then the GUI opens and all the buttons work including the add button which adds items using JOptionPane with included try catch statements to verify input. Added try catch around all methods using files to ensure that if we do implement user file entry they will not crash the program but rather will just send a error message to the consol saying what went wrong and then continue forward. Will still need to add a if statement to the add method in the datamanager to ensure that if the file cant be found the item wont be added to the multilist, but again even if that happens the program wont crash it just wont write to the file.