GUI Update and Cancel Item Add
Redid the GUI class using GridBagLayout for a more dynamic layout, since previous layout would warp when resizing the screen. It's not crazy fancy but it does work 100%, also made it viable for the user to press cancel during the JOptionPane input screens so that they can stop the item adding. Previously this would be caught by the try catch and just force them to accept in an infinite loop. Overall, should be in 100% working order with no known problems and acceptable for submission in its current state. We should look over the code and test the program some to ensure that the system is not weak to any type of cicumstance that I did not forsee ocurring.